Update to move to Preview folder
Moved new controllers and display channels to preview folder, created a helper to keep things DRY in the controller implementations.
