+++ src/Foundation/Features/Preview/CatalogPartialPreviewController.cs
@@ -1,22 +1,23 @@
+﻿using System.Web.Mvc;
+using EPiServer.Commerce.Catalog.ContentTypes;
 using EPiServer.Core;
 using EPiServer.Framework.DataAnnotations;
-using EPiServer.Framework.Web;
-using EPiServer.Web;
+using EPiServer.Framework.Web.Mvc;
 using EPiServer.Web.Mvc;
-using System.Web.Mvc;
 
 namespace Foundation.Features.Preview
 {
     [TemplateDescriptor(
         Inherited = true,
-        TemplateTypeCategory = TemplateTypeCategories.MvcController, //Required as controllers for blocks are registered as MvcPartialController by default
-        Tags = new[] { RenderingTags.Preview, RenderingTags.Edit },
+        Tags = new[] { PartialViewDisplayChannel.PartialViewDisplayChannelName },
         AvailableWithoutTag = false)]
-    public class PreviewController : ActionControllerBase, IRenderTemplate<BlockData>
+    [VisitorGroupImpersonation]
+    [RequireClientResources]
+    public class CatalogPartialPreviewController : ContentController<CatalogContentBase>
     {
         private readonly PreviewControllerHelper _previewControllerHelper;
 
-        public PreviewController(PreviewControllerHelper previewControllerHelper)
+        public CatalogPartialPreviewController(PreviewControllerHelper previewControllerHelper)
         {
             _previewControllerHelper = previewControllerHelper;
         }
